add hoc with-screen-switch
