+++ package.json
@@ -30,10 +30,13 @@
     "node": "10"
   },
   "dependencies": {
+    "@types/jest": "^24.0.13",
     "prop-types": "15.7.2",
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "react-redux": "7.0.3",
-    "redux": "4.0.1"
+    "recompose": "^0.30.0",
+    "redux": "4.0.1",
+    "why-did-you-update": "^1.0.6"
   }
 }
